backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
flask==3.0.0
flask-cors==4.0.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic>=2.9.0,<3.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
scikit-learn==1.3.2
websockets>=13.0.0,<15.1.0
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
xgboost==2.0.3
joblib==1.3.2
anyio>=4.8.0,<5.0.0
# Note: ml-dtypes and tensorboard versions may conflict with tensorflow/jax if installed
# These are optional dependencies for our project
ml-dtypes>=0.4.0,<0.5.0
tensorboard>=2.18,<2.19
# Jupyter Notebook
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.6

